Moved config methods in to separate module
