+++ lib/ztd/todo/supervisor.ex
@@ -1,5 +1,7 @@
 defmodule ZTD.Todo.Supervisor do
   use Supervisor
+
+  alias  ZTD.Todo.Config
   import Supervisor.Spec
 
 
@@ -27,28 +29,6 @@ defmodule ZTD.Todo.Supervisor do
   end
 
 
-  @doc "Return the Application mode"
-  def mode do
-    mode =
-      :env_var
-      |> config()
-      |> System.get_env
-      |> normalize()
-
-    case (mode in allowed_modes()) do
-      true  -> mode
-      false -> raise "Unknown Application Mode"
-    end
-  end
-
-
-  @doc "Return the correct adapter for config"
-  def adapter do
-    config(:adapters) |> Keyword.get(mode())
-  end
-
-
-
 
 
 
@@ -57,7 +37,7 @@ defmodule ZTD.Todo.Supervisor do
 
 
   def init(:ok) do
-    mode()
+    Config.mode()
     |> children()
     |> Supervisor.init(strategy: :one_for_one)
   end
@@ -66,41 +46,21 @@ defmodule ZTD.Todo.Supervisor do
 
 
 
-  ## Private Helpers
-  ## ---------------
+  ## Children Spec
+  ## -------------
 
 
-  # Define children for application modes
+  # Children for Engine Mode
   defp children(:engine) do
     [
       supervisor(ZTD.Repo, []),
     ]
   end
 
+
+  # Raise error for other modes
   defp children(_mode) do
     raise "Supervision Tree not defined for specified mode"
-  end
-
-
-  # Get Config
-  defp config do
-    Application.get_env(:ztd, :mode)
-  end
-
-  defp config(key, default \\ nil) when is_atom(key) do
-    Keyword.get(config(), key, default)
-  end
-
-
-  defp allowed_modes do
-    Keyword.keys(config(:adapters))
-  end
-
-
-  defp normalize(str) do
-    "#{str}"
-    |> String.downcase
-    |> String.to_existing_atom
   end
